server/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2

sqlalchemy==2.0.41
psycopg2==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1

scikit-learn==1.6.1
joblib==1.5.0
pandas==2.2.3
keras==3.9.2

python-jose==3.4.0
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
email_validator==2.2.0

python-dotenv==1.1.0